requirements.txt
streamlit==1.10.0
dotenv
google-generativeai==0.6.0
googletrans==3.1.0a  # Stable version of googletrans
httpx  # Ensure compatibility with newer googletrans versions
gtts==2.2.3
youtube-transcript-api==0.4.4
graphviz==0.20.1
re==2.2.1
random==2.3



